requirements.txt
gradio
numpy
Pillow
colour-science
scipy
piexif
psutil
